refactor(tunnel): split HA runtime into tunnel crate
