+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
   "lib/concourse-client",
   "lib/zenduty-client",
   "core",
+  "core/crates/tunnel",
   "core/crates/library",
   "core/crates/tool-caching",
   "mcp-gateway",
@@ -64,6 +65,7 @@ openai-client = { path = "lib/openai-client" }
 github-app = { path = "lib/github-app" }
 fake-mcp-upstream = { path = "lib/fake-mcp-upstream" }
 drua-git-proxy = { path = "lib/git-proxy" }
+drua-tunnel = { path = "core/crates/tunnel" }
 
 # MCP
 rmcp = { version = "1.2", features = ["server", "client", "transport-streamable-http-server", "transport-streamable-http-client", "transport-streamable-http-client-reqwest", "macros"] }
