+++ core/crates/tunnel/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "drua-tunnel"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+async-trait = "0.1"
+chrono = { version = "0.4", features = ["clock", "serde"], default-features = false }
+rmcp = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+sqlx = { version = "0.8", default-features = false, features = ["runtime-tokio-rustls", "postgres", "uuid", "chrono", "json"] }
+thiserror = { workspace = true }
+tokio = { version = "1", features = ["rt", "sync", "time"] }
+tracing = { workspace = true }
+uuid = { version = "1", features = ["serde", "v4"] }
